board_inspector: fix PCI BAR bitfield layout on Python 3.13+/3.14
MemoryBar32/MemoryBar64/IOBar declared their leading flag bits as c_uint8
bitfields followed by a wide c_uint32/c_uint64 bitfield for the base
address. Python 3.13 changed ctypes so bitfields of different storage
types no longer share a storage unit, so each struct grew by one byte
(5/9/5 instead of 4/8/4). That extra byte shifted every field after the
BARs in the PCI type-1 header, making secondary_bus_number read as 0.

The visible symptom: board XMLs generated on Ubuntu 24.04+/26.04 had
every PCI bus at address 0x0, which later breaks static allocation
("multiple element nodes are found for //bus[...]/device[...]").

Use a single storage type for all bitfields in each BAR struct so they
pack into one 32-/64-bit unit as intended. Verified sizes are back to
4/8/4 and secondary_bus_number now decodes correctly.

+++ misc/config_tools/board_inspector/pcieparser/header.py
@@ -25,10 +25,14 @@ class Common(cdata.Struct):
 
 class MemoryBar32(cdata.Struct):
     _pack_ = 1
+    # All bitfields must share the same storage type (c_uint32). Python 3.14
+    # changed ctypes bitfield packing so mixed-type bitfields no longer share a
+    # storage unit, which would make this struct 5 bytes instead of 4 and
+    # misalign every following field in the PCI header.
     _fields_ = [
-        ('indicator', ctypes.c_uint8, 1),
-        ('type', ctypes.c_uint8, 2),
-        ('prefetchable', ctypes.c_uint8, 1),
+        ('indicator', ctypes.c_uint32, 1),
+        ('type', ctypes.c_uint32, 2),
+        ('prefetchable', ctypes.c_uint32, 1),
         ('base_z', ctypes.c_uint32, 28),
     ]
 
@@ -40,10 +44,11 @@ class MemoryBar32(cdata.Struct):
 
 class MemoryBar64(cdata.Struct):
     _pack_ = 1
+    # All bitfields must share the same storage type (c_uint64); see MemoryBar32.
     _fields_ = [
-        ('indicator', ctypes.c_uint8, 1),
-        ('type', ctypes.c_uint8, 2),
-        ('prefetchable', ctypes.c_uint8, 1),
+        ('indicator', ctypes.c_uint64, 1),
+        ('type', ctypes.c_uint64, 2),
+        ('prefetchable', ctypes.c_uint64, 1),
         ('base_z', ctypes.c_uint64, 60),
     ]
 
@@ -55,9 +60,10 @@ class MemoryBar64(cdata.Struct):
 
 class IOBar(cdata.Struct):
     _pack_ = 1
+    # All bitfields must share the same storage type (c_uint32); see MemoryBar32.
     _fields_ = [
-        ('indicator', ctypes.c_uint8, 1),
-        ('reserved', ctypes.c_uint8, 1),
+        ('indicator', ctypes.c_uint32, 1),
+        ('reserved', ctypes.c_uint32, 1),
         ('base_z', ctypes.c_uint32, 30),
     ]
